Initialize A Basic Shell Project
Created a shell program that is written in C and built using CMake. In
the current state it only has the basic functionality for minimal
usefulness. The current project plan is included in the README.

Next step will be to add basic arguments to the shell and allow for
redirection of input and output. Once that is completed the next main
feature will be adding the background processes with the additional
signal handlers for proper functionality.

The future steps are other features that will improve the shell, but not
add large fundamental features. This could change if new ideas come to
mind for the project.

Signed-off-by: Jacob Evans <jacob@jwevans.dev>