Added Command Line Arguments
Add arguments to the command line that can display help messages and
version/copyright information. Apart from these arguments ther was also
added an argument for a command `-c` in which the shell executes that
command then exits. Additionally, and option for controllling the
verbosity of the messages was added, but still needs work fully be at a
working position.

The next step is to add input and output redirection to the shell which
will allow for proper handling of background processes. After that has
been added I will pause the further development of features to cleanup
the code, add comments, and add tests.

Signed-off-by: Jacob Evans <jacob@jwevans.dev>