Update Parsing and Organization
This changed the parsing to allow for better string arguments and for
the inclusion of environment variables. In this change some features
have been temporarily removed. This primarily includes the execution of
commands and builtins.

Once this reaches parity with the current main these changes wil be
merged into the main branch.

Signed-off-by: Jacob Evans <jacob@jwevans.dev>