Add Background Process and Redirection
Allows background processes with the `&` as the last non-whitespace or
comment character. These will be harvested upon a new prompt being
shown.

The redirection is only for files and either to stdin or from stdout.
This uses the `<` for input and `>` for output. These can occur at any
point within the command string.

To facilitate these signal handlers have also been added meaning that
`Ctrl+C` can no longer be used to exit the shell. Similarly `Ctrl+Z`
will not background the process.

Signed-off-by: Jacob Evans <jacob@jwevans.dev>